requirements.txt
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
alembic==1.14.0
fastapi==0.115.6
pydantic==2.10.5
uvicorn==0.34.0
PyJWT==2.10.1
pytest==8.3.4
httpx==0.28.1